+++ bzt/engine.py
@@ -66,13 +66,15 @@ class Engine(object):
         self.log = parent_logger.getChild(self.__class__.__name__)
         self.config = Configuration()
         self.config.log = self.log.getChild(Configuration.__name__)
-        self.modules = {}
+        self.modules = {}  # available modules
         self.provisioning = Provisioning()
         self.aggregator = EngineModule()  # FIXME: have issues with non-aggregator object set here
         self.interrupted = False
         self.check_interval = 1
         self.stopping_reason = None
         self.engine_loop_utilization = 0
+        self.prepared = []
+        self.started = []
 
     def configure(self, user_configs, read_config_files=True):
         """
@@ -117,9 +119,18 @@ class Engine(object):
             self.__prepare_provisioning()
             self.__prepare_reporters()
             self.config.dump()
+
         except BaseException as exc:
             self.stopping_reason = exc if not self.stopping_reason else self.stopping_reason
             raise
+
+    def _startup(self):
+        modules = self.services + [self.aggregator] + self.reporters + [self.provisioning]
+        for module in modules:
+            self.log.debug("Startup %s", module)
+            self.started.append(module)
+            module.startup()
+        self.config.dump()
 
     def run(self):
         """
@@ -139,26 +150,23 @@ class Engine(object):
         finally:
             self._shutdown()
 
-    def _startup(self):
-        modules = self.services + [self.aggregator] + self.reporters + [self.provisioning]
-        for _module in modules:
-            _module.startup()
-        self.config.dump()
+    def _check_modules_list(self):
+        finished = False
+        modules = [self.provisioning, self.aggregator] + self.services + self.reporters
+        for module in modules:
+            if module in self.started:
+                self.log.debug("Checking %s", module)
+                finished |= module.check()
+        return finished
 
     def _wait(self):
         """
         Wait modules for finish
         :return:
         """
-        self.log.info("Waiting for finish...")
         prev = time.time()
-        modules = []
-        if self.provisioning:
-            modules.append(self.provisioning)
-        if self.aggregator:
-            modules.append(self.aggregator)
-        modules += self.services + self.reporters
-        while not EngineModule.check_modules_list(modules):
+
+        while not self._check_modules_list():
             now = time.time()
             diff = now - prev
             delay = self.check_interval - diff
@@ -178,12 +186,11 @@ class Engine(object):
         """
         self.log.info("Shutting down...")
         exception = None
-        modules = [self.provisioning, self.aggregator]
-        modules += self.reporters
-        modules += self.services
+        modules = [self.provisioning, self.aggregator] + self.reporters + self.services
         for module in modules:
             try:
-                module.shutdown()
+                if module in self.started:
+                    module.shutdown()
             except BaseException as exc:
                 self.log.error("Error while shutting down: %s", traceback.format_exc())
                 self.stopping_reason = exc if not self.stopping_reason else self.stopping_reason
@@ -201,11 +208,11 @@ class Engine(object):
         self.log.info("Post-processing...")
         # :type exception: BaseException
         exception = None
-        modules = [self.provisioning, self.aggregator] + self.reporters
-        modules += self.services
+        modules = [self.provisioning, self.aggregator] + self.reporters + self.services
         for module in modules:
             try:
-                module.post_process()
+                if module in self.prepared:
+                    module.post_process()
             except KeyboardInterrupt as exc:
                 self.log.error("Shutdown: %s", exc)
                 self.stopping_reason = exc if not self.stopping_reason else self.stopping_reason
@@ -428,6 +435,7 @@ class Engine(object):
         if not cls:
             raise ValueError("Please configure provisioning settings")
         self.provisioning = self.instantiate_module(cls)
+        self.prepared.append(self.provisioning)
         self.provisioning.prepare()
 
     def __prepare_reporters(self):
@@ -445,6 +453,7 @@ class Engine(object):
 
         # prepare reporters
         for module in self.reporters:
+            self.prepared.append(module)
             module.prepare()
 
     def __prepare_services(self):
@@ -461,6 +470,7 @@ class Engine(object):
             self.services.append(instance)
 
         for module in self.services:
+            self.prepared.append(module)
             module.prepare()
 
     def __prepare_aggregator(self):
@@ -474,6 +484,7 @@ class Engine(object):
             self.aggregator = EngineModule()
         else:
             self.aggregator = self.instantiate_module(cls)
+        self.prepared.append(self.aggregator)
         self.aggregator.prepare()
 
     def _set_up_proxy(self):
@@ -662,6 +673,8 @@ class EngineModule(object):
         self.engine = None
         self.settings = BetterDict()
         self.parameters = BetterDict()
+        self.delay = 0
+        self.start_time = None
 
     def prepare(self):
         """
@@ -701,24 +714,6 @@ class EngineModule(object):
         Do all possibly long analysis and processing on run results
         """
         pass
-
-    @staticmethod
-    def check_modules_list(modules, require_all=False):
-        """
-        Helper for bulk check
-
-        :type modules: list
-        :param require_all:
-        :return:
-        """
-        finished = require_all
-        for module in modules:
-            logging.debug("Checking %s", module)
-            if require_all:
-                finished &= module.check()
-            else:
-                finished |= module.check()
-        return finished
 
 
 class Provisioning(EngineModule):