+++ site/dat/docs/Changelog.md
@@ -7,6 +7,7 @@
 
 ## 1.3.3 <sup>24 mar 2016</sup>
  - add new `hostaliases` setting for all executors
+ - add delay capability to engine 
 
 ## 1.3.2 <sup>23 mar 2016</sup>
  - fix lowercase hostname for JMeter HTTP request